Send namespace in diagnostic message type (#1)
* Add namespace prefixes to diagnostic status type field.

Also append namespace information to human readable messages to assist in triage. The namespace can be configured as 'com-1', 'com-2', etc. so we can display which computer is malfunctioning.

* Can't display which mount point is in error.

Diagnostic messages get overwritten by valid mount points right now with the highest error level even if that disk is not in error.

* If no namespace specified use the hostname.